+++ packages/orq-rc/RELEASES.md
@@ -10869,3 +10869,13 @@ Based on:
 - [python v4.15.0-rc.40] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.40 - packages/orq-rc
+
+## 2026-09-11 14:46:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.41] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.41 - packages/orq-rc